+++ tests/benchmarks/comparison.sh
@@ -1,4 +1,6 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+cd "$(dirname "${BASH_SOURCE[0]}")"
 
 if ! which hyperfine > /dev/null 2>&1; then
     echo "'hyperfine' does not seem to be installed."